+++ demo/admin/src/products/future/ProductsGrid.cometGen.tsx
@@ -93,7 +93,15 @@ export default defineConfig<GQLProduct>({
         },
         { type: "text", name: "description", headerName: "Description" },
         // TODO: Allow setting options for `intl.formatNumber` through `valueFormatter` (type "number")
-        { type: "number", name: "price", headerName: "Price", maxWidth: 150, headerInfoTooltip: "Price in EUR", visible: "up('md')" },
+        {
+            type: "number",
+            name: "price",
+            currency: "EUR",
+            headerName: "Price",
+            maxWidth: 150,
+            headerInfoTooltip: "Price in EUR",
+            visible: "up('md')",
+        },
         {
             // TODO: Implement showing actual label in `valueFormatter` (type "staticSelect")
             type: "staticSelect",
@@ -131,6 +139,13 @@ export default defineConfig<GQLProduct>({
             headerName: "Manufacturer",
             fieldName: "manufacturer",
             filterOperators: ManufacturerFilterOperators,
+        },
+        {
+            type: "virtual",
+            name: "tags",
+            headerName: "Tags",
+            queryFields: ["tags.title"],
+            renderCell: ({ row }) => <>{row.tags.map((tag) => tag.title).join(", ")}</>,
         },
         {
             type: "actions",