+++ packages/admin/admin-color-picker/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @comet/admin-color-picker
 
+## 8.0.0-beta.4
+
+### Patch Changes
+
+- Updated dependencies [a93455f]
+- Updated dependencies [72d1a5e]
+- Updated dependencies [1c62e87]
+    - @comet/admin@8.0.0-beta.4
+    - @comet/admin-icons@8.0.0-beta.4
+
 ## 8.0.0-beta.3
 
 ### Patch Changes
@@ -76,6 +86,21 @@
 - Updated dependencies [682a674]
     - @comet/admin@8.0.0-beta.0
     - @comet/admin-icons@8.0.0-beta.0
+
+## 7.19.0
+
+### Patch Changes
+
+- Updated dependencies [3544127ad]
+    - @comet/admin@7.19.0
+    - @comet/admin-icons@7.19.0
+
+## 7.18.0
+
+### Patch Changes
+
+- @comet/admin@7.18.0
+- @comet/admin-icons@7.18.0
 
 ## 7.17.0