+++ packages/admin/admin-generator/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @comet/admin-generator
 
+## 8.0.0-beta.4
+
 ## 8.0.0-beta.3
 
 ## 8.0.0-beta.2